Add remaining javadocs and extract Maze as interface
